requirements.txt
librosa
tensorflow
tensorflow_probability
tensorflow_hub
tensorflow_io
numpy
python-dateutil
matplotlib
psutil
audiomentations
ydf
pandas
